requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
Deprecated==1.2.13
Django==4.1
django-debug-toolbar==3.5.0
django-filter==22.1
djangorestframework==3.13.1
filelock==3.7.1
hyperlink==21.0.0
icrawler==0.6.6
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.1
Markdown==3.4.1
packaging==21.3
parsel==1.6.0
Pillow==9.2.0
prompt-toolkit==3.0.30
Protego==0.2.1
proxyscrape==0.3.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dotenv==0.20.0
python-scrapyd-api==2.1.2
pytz==2022.1
queuelib==1.6.2
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
rq==1.11.0
Scrapy==2.6.2
scrapy-djangoitem==1.1.1
scrapy-proxy-pool==0.1.9
scrapy-user-agents==0.1.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tldextract==3.3.1
Twisted==22.4.0
typing_extensions==4.3.0
ua-parser==0.15.0
urllib3==1.26.11
user-agents==2.2.0
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.4.0
